P10: Judicial Committee
P10: Judicial Committee

AKA P9 in "2021 Bylaws Committee Report" PDF on lpid.org

Original Section: CONSTITUTION ARTICLE VII: COMMITTEES SECTION 3

Revised Section: ARTICLE VII: OTHER COMMITTEES - SECTION 8

Purpose: To expand on the definitions and duties of the legislative
committee and to bring the required number of members down to match
national standards.

Proposal Notes: The word "yearly" is intended to be adjusted if Proposal
2 - Frequency of the Regular Convention is approved.

Motion: Jennifer Imhoff moved with second to vote yes. The motion passed
with no objections on 3/27/2021

Preparation Notes: The text this modifies differs trivially between the
original documents and the version from Proposal 1, and I changed the
word "Committee" to "Board" in the added paragraph. --Zach