P4: Quorum for Regular Conventions
P4: Quorum for Regular Conventions

Original Section: CONSTITUTION - ARTICLE VI: CONVENTIONS - new section

Revised Section: ARTICLE V: MEETINGS - new section

Purpose: As we have seen, quorum for Regular Conventions has been an
issue due to the current rigidness of our bylaws and Robert’s Rules of
Order. Not only is this an issue if a physical convention cannot be
held, but it also poses a recurring issue to those members who are
simply too far away to participate in a meeting such as a Regular
Convention. We have no guarantee that similar circumstances will not
arise in the future. Setting a consistent threshold for physical,
hybrid, and online Regular Conventions is imperative to keep the party
functioning in the future.

Motion: Jennifer Imhoff moved with second to vote yes. The motion passed
with no objections on 1/23/2021