P3: Staggering Elec. of Off. and Exec. Board Memb.
P3: Staggering Election of Officers and Executive Board Members

Original Section:  CONSTITUTION - ARTICLE IV: CONVENTION - SECTION 1

Revised Section: ARTICLE IV: OFFICERS - SECTION 1

Purpose: Currently, should proposal number 1 be passed, elections for the
officers of the party as well as the executive committee would be held every
year, effectively cutting the term of each of these positions from two years to
one year. This proposal would remedy this by allowing for elections of Officers
in even numbered years, and the election of other Executive Committee members
in odd numbered years, This would keep the importance of each annual convention
whole, and allow the individual components of the Executive committee to be
periodically checked without the chance of a complete overhaul of leadership in
a single convention.

Motion: Jennifer Imhoff moved with second to vote yes. The motion
passed with no objections on 1/23/2021

Preparation Notes: In Proposal 1, the sentence beginning "If, for any
reason, a Regular Convention is not held..." was removed, and I've
re-added here. Compared to the official committee report, I also added a
comma and uncapitalized the mid-sentence occurrence of "Shall". --Zach