P6: Regional Reps. to be Elec. via Region. Caucus
P6: Regional Representatives to be Elected via Regional Caucus

Original Section: CONSTITUTION - ARTICLE V: EXECUTIVE COMMITTEE -
SECTION 1

Revised Section: ARTICLE VI: EXECUTIVE BOARD - SECTION 1

Purpose: Generalize the description to allow for natural party growth
without requiring additional bylaw changes that do not impact the spirit
of this section.

Motion: Robert Imhoff moved with second to vote yes. The motion passed
with no objections on 1/23/2021

Preparation Notes: This change differs from how it's displayed in the
committee report because the first sentence shown in "Original" text
differs from both the original documents and from the version created by
Proposal 1. In the actual original, the section reads, in its entirety,
"The Executive Committee of the Party shall be composed of the
following: The four elected officers of the Party." In Proposal 1, it
reads, in its entirety, "The officers of the Party and three Regional
Representatives shall constitute the Executive Board."  I re-added text
very similar to that in the actual original documents, substituting
"Committee" for "Board". Additionally, the text beginning, "Each
Regional Chair shall...", was removed in Proposal 1, so I've re-added it
here, in its amended form. --Zach