P2: Frequency of the Regular Convention
P2: Frequency of the Regular Convention

Original Section: CONSTITUTION - ARTICLE VI: CONVENTION - SECTION 1

Revised Section: ARTICLE V: MEETING - SECTION 1

Purpose: The Regular Convention is currently held at an interval of 2
years, with a convention occurring on every even numbered year, per the
LPID Constitution. Increasing the frequency of the Regular Conventions
will allow the State Party to tweak and adjust the bylaws and
constitution to prevent unforeseen problems in the wording of the bylaws
or constitution from becoming problematic for an extended period of
time. It also allows party members greater access to the State Party and
encourages new recruits to become involved on the state level to help
grow our party.

Moving to April instead of June will get LPID internal business out of
the way earlier in the year and allow the parties members and leadership
to shift focus to our candidates and offer what help is deemed
appropriate to their campaigns.

Motion: Dan Karlan moved with second to vote yes. The motion passed with
no objections on 1/23/2021