P12: Transaction of Business by E-mail
P12: Transaction of Business by E-mail

AKA P11 in "2021 Bylaws Committee Report" PDF on lpid.org

Original Section: BYLAWS ARTICLE VI: EXECUTIVE COMMITTEE - Section 3

Revised Section: ARTICLE VI: EXECUTIVE BOARD - Section 5

Purpose: The Executive Board may, without meeting together, transact business
by mail and/or e-mail, voting on a question submitted to them by or with the
approval of the Chair or co-sponsored by three (3) members of the Executive
Board. Ten days shall be allowed for the return of votes thereon by e-mail to
the Party Secretary during the period between a Regular Convention and the next
general election, and twenty days at all other times.

The result of this change is to move the LPID out of the 20th century and fully
into the 21st century. Use of mail, when e-mail is available, is an
inappropriate holdover from days when it could not be counted on for all Board
members to have e-mail. It further introduces the complication that all mail
into and out of Pocatello (for example) now is directed through Salt Lake City
– including mail from one Pocatello address to another address within Pocatello
(I can prove this) – resulting in unnecessary delays, especially subject to
local weather conditions outside of Idaho, and placing the Bylaws-specified
cutoff time of 10 days in jeopardy of being missed. Now that all Board members
are expected to have e-mail access, even through smartphones, continuing to
refer to the mid-20th-century days when e-mail was erratic is inappropriate.

Motion: Dan Karlan moved with second to vote yes. The motion passed with no
objections on 3/27/2021

Preparation Notes: The "Original" text in the committee report reflects
the correction of the "voting a question" typographical error, present
in both the original documents and the version from Proposal 1. --Zach